Refactor for clarity, dead code removal, and version bump.
