+++ src/CMakeLists.txt
@@ -1,9 +1,9 @@
 project (OSL)
 
 # Release version of the library
-set (OSL_LIBRARY_VERSION_MAJOR 0)
-set (OSL_LIBRARY_VERSION_MINOR 6)
-set (OSL_LIBRARY_VERSION_PATCH 2)
+set (OSL_LIBRARY_VERSION_MAJOR 1)
+set (OSL_LIBRARY_VERSION_MINOR 0)
+set (OSL_LIBRARY_VERSION_PATCH 0)
 
 # Version of the OSO file format and instruction set
 set (OSO_FILE_VERSION_MAJOR 1)
