+++ src/doc/languagespec.tex
@@ -44,7 +44,7 @@
 
 \def\langname{Open Shading Language\xspace}
 \def\product{{\sffamily Open Shading Language}\xspace}
-\def\versionnumber{0.6}
+\def\versionnumber{1.0}
 \def\productver{\product\ {\sffamily \versionnumber}\xspace}
 
 
@@ -63,8 +63,8 @@ Editor: Larry Gritz \\
 \emph{lg@imageworks.com}
  \bigskip \\
 }
-\date{{\large Date: 10 May, 2011 \\
-(with corrections, 29 September, 2011)
+\date{{\large Date: 12 Oct, 2011 \\
+%(with corrections, 29 September, 2011)
 }}
