+++ src/include/CMakeLists.txt
@@ -1,4 +1,5 @@
-set (public_headers accum.h export.h genclosure.h optautomata.h oslclosure.h
+set (public_headers accum.h dual.h dual_vec.h
+                    export.h genclosure.h optautomata.h oslclosure.h
                     oslcomp.h oslconfig.h oslexec.h oslquery.h )
 message (STATUS "Create oslversion.h from oslversion.h.in")
 configure_file (oslversion.h.in ${CMAKE_BINARY_DIR}/include/oslversion.h @ONLY)
