+++ src/include/dual.h
