+++ src/include/dual_vec.h
