+++ src/include/genclosure.h
@@ -86,6 +86,17 @@ struct ClosureParam {
 
 #define CLOSURE_FINISH_PARAM(st) { TypeDesc(), sizeof(st), NULL, 0 }
 
+
+// Not meant to be public, but there's no better place at the moment.
+// FIXME someday.
+#define CLOSURE_PREPARE(name, classname)    \
+void name(RendererServices *, int id, void *data) \
+{                                                 \
+    memset(data, 0, sizeof(classname));           \
+    new (data) classname();                       \
+}
+
+
 }; // namespace OSL
 
 #ifdef OSL_NAMESPACE
