+++ src/liboslexec/CMakeLists.txt
@@ -16,14 +16,11 @@ SET ( liboslexec_srcs
           rendservices.cpp runtimeoptimize.cpp typespec.cpp
           lpexp.cpp lpeparse.cpp automata.cpp accum.cpp
           opclosure.cpp builtin_closures.cpp
+          llvm_instance.cpp llvm_util.cpp
           ../liboslcomp/ast.cpp ../liboslcomp/codegen.cpp
           ../liboslcomp/oslcomp.cpp ../liboslcomp/symtab.cpp
           ../liboslcomp/typecheck.cpp
     )
-
-if (LLVM_FOUND)
-  SET ( liboslexec_srcs ${liboslexec_srcs} llvm_instance.cpp )
-endif ()
 
 FILE ( GLOB compiler_headers "*.h" )
