+++ src/liboslexec/background.cpp
@@ -26,10 +26,8 @@ THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-#include <OpenImageIO/fmath.h>
-
+#include "oslexec_pvt.h"
 #include "genclosure.h"
-#include "oslops.h"
 #include "oslclosure.h"
