+++ src/liboslexec/bsdf_cloth.cpp
@@ -32,9 +32,8 @@ OF THIS SOFTWARE, even IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <OpenEXR/ImathFun.h>
 
 #include "genclosure.h"
-#include "oslops.h"
 #include "oslexec_pvt.h"
-
+#include "dual.h"
 #include "bsdf_cloth_fncs.h"
 
 inline bool odd(int x)  { return (x & 1) == 1; }
