+++ src/liboslexec/bsdf_cloth_fncs.h
@@ -28,7 +28,6 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 
 
-#include "oslops.h"
 #include "oslexec_pvt.h"
 #include <OpenEXR/ImathMatrix.h>
