+++ src/liboslexec/bsdf_diffuse.cpp
@@ -28,7 +28,6 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include <OpenImageIO/fmath.h>
 
-#include "oslops.h"
 #include "genclosure.h"
 #include "oslexec_pvt.h"
