+++ src/liboslexec/bsdf_fakefur.cpp
@@ -29,7 +29,6 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <OpenImageIO/fmath.h>
 
 #include "genclosure.h"
-#include "oslops.h"
 #include "oslexec_pvt.h"
 
 // based on the implementation by Dan B. Goldman
