+++ src/liboslexec/context.cpp
@@ -39,7 +39,6 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <OpenImageIO/thread.h>
 
 #include "oslexec_pvt.h"
-#include "oslops.h"
 
 #ifdef OSL_NAMESPACE
 namespace OSL_NAMESPACE {
