+++ src/liboslexec/master.cpp
@@ -39,7 +39,6 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "OpenImageIO/thread.h"
 
 #include "oslexec_pvt.h"
-#include "oslops.h"
 #include "../liboslcomp/oslcomp_pvt.h"
