+++ src/liboslexec/opclosure.cpp
@@ -28,7 +28,6 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include <cmath>
 
-#include "oslops.h"
 #include "oslexec_pvt.h"
 #include "genclosure.h"
