+++ src/liboslexec/opcloud.cpp
@@ -26,7 +26,6 @@ THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-#include "oslops.h"
 #include "oslexec_pvt.h"
 
 inline ustring USTR(const char *cstr) { return (*((const ustring *)&cstr)); }
