+++ src/liboslexec/opnoise.cpp
@@ -29,7 +29,6 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <limits>
 
 #include "oslexec_pvt.h"
-#include "oslops.h"
 #include "noiseimpl.h"
