+++ src/liboslexec/oslexec_pvt.h
@@ -1174,6 +1174,7 @@ tex_interp_to_code (ustring modename)
 }
 
 
+
 }; // namespace OSL
 
 #ifdef OSL_NAMESPACE
