+++ src/liboslexec/vol_subsurface.cpp
@@ -28,7 +28,6 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include <OpenImageIO/fmath.h>
 
-#include "oslops.h"
 #include "oslexec_pvt.h"
 #include "genclosure.h"
